LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-button-disabled-no-focus-expected.txt


FAIL CSS Test: disabled ::scroll-button() shouldn't be focusable assert_equals: expected "rgb(0, 128, 0)" but got ""
